+++ Source/Scripts/EasternKingdoms/Karazhan/Nightbane.cs
@@ -137,7 +137,7 @@ namespace Scripts.EasternKingdoms.Karazhan.Nightbane
             SetupGroundPhase();
         }
 
-        public override void DamageTaken(Unit attacker, ref float damage, DamageEffectType damageType, SpellInfo spellInfo = null)
+        public override void DamageTaken(Unit attacker, ref double damage, DamageEffectType damageType, SpellInfo spellInfo = null)
         {
             if (phase == NightbanePhases.Fly)
             {
