+++ Source/Scripts/Spells/DemonHunter/spell_demon_hunter_pain.cs
@@ -30,7 +30,7 @@ public class spell_demon_hunter_pain : AuraScript, IHasAuraEffects
 		if (damageTaken <= 0)
 			return;
 
-		var painAmount = (50.0f * (float)damageTaken) / (float)caster.GetMaxHealth();
+		var painAmount = (50.0f * (double)damageTaken) / (double)caster.GetMaxHealth();
 		caster.CastSpell(caster, DemonHunterSpells.REWARD_PAIN, (int)painAmount);
 	}
